+++ src/components/pages/auth/interface.ts
